+++ ShipmentGeek/Properties/Resources.Designer.cs
@@ -63,9 +63,9 @@ namespace ShipmentGeek.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
-        internal static System.Drawing.Bitmap clean {
+        internal static System.Drawing.Bitmap add {
             get {
-                object obj = ResourceManager.GetObject("clean", resourceCulture);
+                object obj = ResourceManager.GetObject("add", resourceCulture);
                 return ((System.Drawing.Bitmap)(obj));
             }
         }